+++ components/pmic/pca9422/fsl_pca9422.h
@@ -999,7 +999,7 @@ typedef enum _pca9422_bx_fpwm_mode
 #define PCA9422_BUCK1_OUT_MAX  (1975000U) /* 1975mV */
 #define PCA9422_BUCK1_OUT_VAL(_volt) \
     (((_volt) - PCA9422_BUCK1_OUT_MIN + PCA9422_BUCK1_OUT_STEP - 1U) / PCA9422_BUCK1_OUT_STEP)
-#define PCA9422_BUCK1_OUT_VOLT(_val) (((_val) * PCA9422_BUCK1_OUT_STEP) + PCA9422_BUCK1_OUT_MIN)
+#define PCA9422_BUCK1_OUT_VOLT(_val) (((uint32_t)(_val) * PCA9422_BUCK1_OUT_STEP) + PCA9422_BUCK1_OUT_MIN)
 
 /*! @brief PCA9422 Buck2 OUT voltage definition. (uV unit) */
 #define PCA9422_BUCK2_OUT_MIN  (400000U)  /* 400mV */
@@ -1007,7 +1007,7 @@ typedef enum _pca9422_bx_fpwm_mode
 #define PCA9422_BUCK2_OUT_MAX  (3400000U) /* 3400mV */
 #define PCA9422_BUCK2_OUT_VAL(_volt) \
     (((_volt) - PCA9422_BUCK2_OUT_MIN + PCA9422_BUCK2_OUT_STEP - 1U) / PCA9422_BUCK2_OUT_STEP)
-#define PCA9422_BUCK2_OUT_VOLT(_val) (((_val) * PCA9422_BUCK2_OUT_STEP) + PCA9422_BUCK2_OUT_MIN)
+#define PCA9422_BUCK2_OUT_VOLT(_val) (((uint32_t)(_val) * PCA9422_BUCK2_OUT_STEP) + PCA9422_BUCK2_OUT_MIN)
 
 /*! @brief PCA9422 Buck3 OUT voltage definition. (uV unit) */
 #define PCA9422_BUCK3_OUT_MIN  (400000U)  /* 400mV */
@@ -1015,7 +1015,7 @@ typedef enum _pca9422_bx_fpwm_mode
 #define PCA9422_BUCK3_OUT_MAX  (1975000U) /* 1975mV */
 #define PCA9422_BUCK3_OUT_VAL(_volt) \
     (((_volt) - PCA9422_BUCK3_OUT_MIN + PCA9422_BUCK3_OUT_STEP - 1U) / PCA9422_BUCK3_OUT_STEP)
-#define PCA9422_BUCK3_OUT_VOLT(_val) (((_val) * PCA9422_BUCK3_OUT_STEP) + PCA9422_BUCK3_OUT_MIN)
+#define PCA9422_BUCK3_OUT_VOLT(_val) (((uint32_t)(_val) * PCA9422_BUCK3_OUT_STEP) + PCA9422_BUCK3_OUT_MIN)
 
 /*! @brief PCA9422 LDO2/3 output capacitor selection definition. */
 typedef enum _pca9422_ldo23_csel
@@ -1056,7 +1056,7 @@ typedef enum _pca9422_ldo23_mode
 #define PCA9422_LDO1_OUT_MAX  (3000000U) /* 3000mV */
 #define PCA9422_LDO1_OUT_VAL(_volt) \
     (((_volt) - PCA9422_LDO1_OUT_MIN + PCA9422_LDO1_OUT_STEP - 1U) / PCA9422_LDO1_OUT_STEP)
-#define PCA9422_LDO1_OUT_VOLT(_val) (((_val) * PCA9422_LDO1_OUT_STEP) + PCA9422_LDO1_OUT_MIN)
+#define PCA9422_LDO1_OUT_VOLT(_val) (((uint32_t)(_val) * PCA9422_LDO1_OUT_STEP) + PCA9422_LDO1_OUT_MIN)
 
 /*! @brief PCA9422 LDO2/3 OUT voltage definition. (uV unit) */
 #define PCA9422_LDO23_OUT_MIN  (500000U)  /* 500mV */
@@ -1064,7 +1064,7 @@ typedef enum _pca9422_ldo23_mode
 #define PCA9422_LDO23_OUT_MAX  (1950000U) /* 1950mV */
 #define PCA9422_LDO23_OUT_VAL(_volt) \
     (((_volt) - PCA9422_LDO23_OUT_MIN + PCA9422_LDO23_OUT_STEP - 1U) / PCA9422_LDO23_OUT_STEP)
-#define PCA9422_LDO23_OUT_VOLT(_val) (((_val) * PCA9422_LDO23_OUT_STEP) + PCA9422_LDO23_OUT_MIN)
+#define PCA9422_LDO23_OUT_VOLT(_val) (((uint32_t)(_val) * PCA9422_LDO23_OUT_STEP) + PCA9422_LDO23_OUT_MIN)
 
 /*! @brief PCA9422 LDO4 OUT voltage definition. (uV unit) */
 #define PCA9422_LDO4_OUT_MIN  (800000U)  /* 800mV */
@@ -1072,7 +1072,7 @@ typedef enum _pca9422_ldo23_mode
 #define PCA9422_LDO4_OUT_MAX  (3300000U) /* 3300mV */
 #define PCA9422_LDO4_OUT_VAL(_volt) \
     (((_volt) - PCA9422_LDO4_OUT_MIN + PCA9422_LDO4_OUT_STEP - 1U) / PCA9422_LDO4_OUT_STEP)
-#define PCA9422_LDO4_OUT_VOLT(_val) (((_val) * PCA9422_LDO4_OUT_STEP) + PCA9422_LDO4_OUT_MIN)
+#define PCA9422_LDO4_OUT_VOLT(_val) (((uint32_t)(_val) * PCA9422_LDO4_OUT_STEP) + PCA9422_LDO4_OUT_MIN)
 
 /*! @brief PCA9422 Buck-Boost FPWM mode definition. */
 typedef enum _pca9422_bb_fpwm_mode
@@ -1149,7 +1149,7 @@ typedef enum _pca9422_reg_lpmode
 #define PCA9422_BB_OUT_STEP       (25000U)   /* 25mV */
 #define PCA9422_BB_OUT_MAX        (5000000U) /* 5000mV */
 #define PCA9422_BB_OUT_VAL(_volt) (((_volt) - PCA9422_BB_OUT_MIN + PCA9422_BB_OUT_STEP - 1U) / PCA9422_BB_OUT_STEP)
-#define PCA9422_BB_OUT_VOLT(_val) (((_val) * PCA9422_BB_OUT_STEP) + PCA9422_BB_OUT_MIN)
+#define PCA9422_BB_OUT_VOLT(_val) (((uint32_t)(_val) * PCA9422_BB_OUT_STEP) + PCA9422_BB_OUT_MIN)
 
 /*! @brief PCA9422 Regulator Voltage definition. */
 typedef enum _pca9422_vout
